vector-backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
flask-migrate
openai
python-dotenv
requests
beautifulsoup4
pydantic
redis
pandas
google-auth
google-auth-oauthlib
readability-lxml
lxml
reportlab
httpx
transformers==4.38.2
sentence-transformers==2.2.2
networkx